+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added edit comments
 - Added edit posts
 - Option to show buttons instead of using swipe actions for comments
+- Added option to save comment to context menu
+- Go directly to user or community in search. More search changes to come.
 
 ## [Version 1.0 (46)] - 2023-10-23T11:20:39Z